commit cloud: add option to empty source workspace post import
Summary:
`hg cloud import` was designed to serve as a one-off migration, however, atm it's being used __every__ single time an OD is reserved/provisioned, leading to the following bug:

if you have commit A in WWW, and then yo run hg cloud improt you'll have A' in fbsource. You keep working, you decide you don't need A' anymore and you hide it. However, when you get a new OD, you'll have A' again cause we reimported it when the new OD was provisioned. For this, we discussed with the WWW people and agreed that best option is to have an option to simply wipe out the source workspace after being imported, that way re-importing becomes a no-op and if people want to work again in WWW they can re-import from fbsource.

Reviewed By: muirdm

Differential Revision: D62449339

fbshipit-source-id: ccdf12a204b23f7023e8877939c43ec3899fc166