mercurial/mutation: stop storing primordial commits
Summary:
We are deprecating the use of primoridal commits.  Currently they are not used, but they are still maintained.

In this diff, we remove generation of primordial commits, and stop fetching them from the database.

This is safe to do, as the previous diff (with the JK flipped) means that any server that doesn't yet have this code will not find the primordial, but also won't search for it, either.

Reviewed By: andreacampi

Differential Revision: D61133825

fbshipit-source-id: 5abffb7b1a591336cdec7dbe7ad87502e96d41fe