manifest: Clarify trait bounds and remove IntermediateLeafId
Summary:
The `Manifest` trait is used to represent types that can function as an async list of entries, where each entry is either one of two associated types, `TreeId` or `LeafId`. Typically `TreeId` is loadable through the generic `Store` type to get something of the same type that implements `Manifest`. On the other hand there's no expectation for `LeafId` to be loadable through the store and in a lot of cases it's not. To clarify this, this diff is renaming `LeafId` to `Leaf` to indicate that leaves don't necessarily have to function as an id.

Additionally this diff removes the `IntermediateLeafId` concept which was introduced and is only used by `Fsnodes`. It's complicating the type bounds in the manifest derivation code and doesn't seem to actually provide any speed benefit.

Splitting this diff would involve as much work as creating it :)

Reviewed By: clara-9

Differential Revision: D62496638

fbshipit-source-id: 9d0bbad765e8851176dfc5aa24bdbd36029ad490