force unmount on ENOTCONN
Summary:
# Context
Eden may have stale mounts when the it is unexpectedly killed and the privhelper is unable to clean up the mounts. When restarting, these can prevent eden from remounting checkouts. The goal of this diff is to identify stale mounts during the startup process and resolve them.

# This Diff
Adds a check isErrorSafeToIgnore on the error returned by stat. If the error is safe to ignore, continue the mount process.
In addition, attempt to unmount the stale mount.
Even if the unmount fails, it is safe to mount over the old mount.

This diff just contains the privhelper changes. this is required to be separate from the integration test because the test fails on sandcastle since the changes it's testing is not on sandcastle's instance of privhelper. I have manually tested this change extensively and can confirm that it works.

# Technical Details
ENOTCONN is the error returned when the mount is disconnected. We use this to determine if the mount is stale or not. If the mount isn't disconnected, it will return either success, or other errors when we try to access it.

# Discussion Points
This diff checks twice for stale mounts during the sanityCheckMountPoint process, once on the mount path and once on a subfile that doesn't exist. The reason for this is that if the kernel has the mount point cached, it may return a valid result for stat even though the mount is hanging. Checking for a file that doesn't exist won't be in the cache so it will return ENOTCONN if the mount is hanging and a file not found error if the mount isn't.

Reviewed By: MichaelCuevas

Differential Revision: D62051804

fbshipit-source-id: fdd8a333b7ceddac31c4db5c76b621ca1afb427e