Split of "[Eden] Add Fix for: CheckoutConfig Corruption" pt3: Add CheckoutConfigCorruption Problem and test
Summary:
# Context
Occasionally, config.toml gets corrupted or deleted. This file stores important information about the eden checkout.

# This Diff
Create the EdenCheckoutConfigCorruption Problem, check for it when trying to read the config, and add test

# Technical Details
- Creates a new CheckoutConfig based on info derived from the CheckoutInfo
- Backs up the old one
- Saves the new CheckoutConfig

# Caveats
This will only work on a currently running mount. If the mount is not running and config.toml is corrupted, we cannot determine what the backing repo is. In this case, we have return a EdenCheckoutInfoCorruption problem instead

Reviewed By: MichaelCuevas

Differential Revision: D62304373

fbshipit-source-id: 96279a165f8738287378de4b7bb3716766518a22