Reduce split modal margins for more working space
Summary:
Make the split and edit stack modals as tall as possible, and more consistent.

This is set up a little weirdly to get the scrolling behavior correct. So we need to adjust some manual pixel offset numbers. Previously, these were tuned for the "Edit stack" modal, but gave too much margin for the single-commit split UI.

Instead, we can use a CSS property so each modal sizing can determine this for itself.

Reviewed By: quark-zju

Differential Revision: D62596521

fbshipit-source-id: 2d1b382a6c10bd287161a52e6e8fa6f9cac60020