case_conflict_skeleton_manifest: implement derivation logic
Summary: Implement derivation logic for `CaseConflictSkeletonManifest`. Since `CCSM` and `BSSM` are both skeleton manifests that transform the paths, derivation of `CCSM` is almost identitcal to `BSSM`. The only two differences are that `CCSM` uses `CssmPath` instead of `BssmPath` to transform the paths, and that `CCSM` derivation doesn't require expanding implicit deletes first. This is because the `CCSM` transformation preserves tree relations between paths i.e. a path is an ancestor of another path in `CCSM` if and only if it was an ancestor of it before the transformation.

Reviewed By: andreacampi

Differential Revision: D62574609

fbshipit-source-id: e1dec9b35f3fa38927af4555faccedad649d1757