Add an async-request submit subcommand
Summary:
## This stack

Convert one (non-Megarepo) SCS method to run via async_requests and the async worker.

## This diff

This is mostly intended for integration tests, but can be useful for manual testing as well.

Short term, it allows me to refactor an ugly test that hardcoded a binary file–ugh. That same test would break the next diff in the stack because of the `monad blobstore upload`, which we can also finally remove.

Reviewed By: clara-9

Differential Revision: D62383425

fbshipit-source-id: 8ce53b9b011caedcadc2ec240095cb56b476cd4d