scm: allow dual backing store build with cas
Summary:
We want to enable CAS in edenfs. The problem is that doing so it makes the test oss binary way bigger, causing regressions in test execution.

For now, the most straightforward solution, according to what was discussed with MichaelCuevas and jdelliot is simply allowing non-cas eden and cas eden to coexist. For this we'll need backing store with and without cas, aswell as sapling backing store with and without cas.

Next diff actually does the sapling changes.

Reviewed By: MichaelCuevas

Differential Revision: D62609935

fbshipit-source-id: e760dc04b227132fc70ce2a432b3238ce508bdf0