Split of "[edenfs] force unmount on ENOTCONN: integration test"
Summary:
# Context
Eden may have stale mounts when the it is unexpectedly killed and the privhelper is unable to clean up the mounts. When restarting, these can prevent eden from remounting checkouts. The goal of this diff is to identify stale mounts during the startup process and resolve them.

# This Diff
This diff is the integration test for automatically unmounting stale mounts on startup. It does this by creating a mount, killing privhelper/eden, checking that the mount is hanging, then restarting eden and checking that the mount is cleared.

# Technical Details
ENOTCONN is the error returned when the mount is disconnected. We use this to determine if the mount is stale or not. If the mount isn't disconnected, it will return either success, or other errors when we try to access it.

# Discussion Points
This diff uses the same method doctor uses to check for stale mounts to determine if a mount is stale/hanging.

Differential Revision: D62472128

fbshipit-source-id: a88c5857881704e3576038db249e2bfef9010319