Allow querying latest state of bonsai_tag_mapping
Summary: The `BonsaiTagMapping` table is used in Git hooks to determine if a tag is annotated or not. We write to this table immediately before invoking the hook so there is a possible race condition if the tag write hasn't been replicated to the DB when the hook gets invoked. The hook would incorrectly assume that the tag is non-annotated and fail the push. This diff avoids that race condition by ensuring that the hook reads from master DB instead of replica so it always gets the latest state.

Reviewed By: YousefSalama

Differential Revision: D62685959

fbshipit-source-id: f01870a8581ee66b4b6a891cf3fbac38c0cfc9b4