Use `cli_log` in git pushrebase binary
Summary:
## This stack
Implements the logic on `git pushrebase` to handle pushes when the source of truth is in the large repo (e.g. fbsource).

## This diff
We should have a structured way to log stuff with verbosity.
I would usually use slog, but there's an effort to stop using it and use `tracing` instead.

I found this library in [the Rust Logging wiki](https://www.internalfb.com/intern/wiki/Rust/Meta-Infra/Logging_in_Rust/). It simplifies logging setup for CLIs.

I'll use it in the next diffs

Reviewed By: andreacampi

Differential Revision: D62740293

fbshipit-source-id: 34af3648feadacdf840a38d1b7725b2566ed065e