use CESID cpp getter
Summary:
# Context
We want to experiment on EdenFS using Meta's mature stack of data analysis and experimentation that is QE + Deltoid. To do this, we need an ID we can join against in data land so we can define a population unit. We're using the [CESID](https://fb.workplace.com/groups/301619822315138/permalink/528321876311597/) for this purpose, which uniquely identifies a dev server, sandcastle instance, ondemand, and laptop/PC.

# This Diff
Diff stack head of changes that add the logging of CESID to EdenFS server. Note that the CESID is already being logged by the edenfsctl cli and we're just catching up on edenFS server side.

# Technical Details
This code leverages 2 things:
- A new C++ implementation of the CESID getter, which is really just file IO.
- Conditional compilation to obscure away Meta-only logic from the OSS project (using a header file introduced in this stack)

# Discussion Points
We had some deliberation prior to finalizing this diff stack. Discussion was basically "Should we use polymorphism/heap allocation or do conditional compilation?".  EdenFS is on the critical path of most tech at Meta, so conditional compilation it is.

Reviewed By: genevievehelsel

Differential Revision: D62485302

fbshipit-source-id: c92b0cd915894c6d5a8e84901a5c5cf8521f2255