add known crawlers to doctor check
Summary:
# Context

We want to know if known crawlers may be the cause of high inode counts on macOS. If so, we can make fixing the known crawlers a higher priority while we wait for the vendor that owns the known crawler to fix their code.

# This diff

This diff adds the concept of known crawlers to the macOS high inode check and puts any known crawlers in the problem description (so that they are logged to Scuba). This will ensure we can correlate any high inode counts with known crawlers being run/installed.

Differential Revision: D62610747

fbshipit-source-id: 085567da901d5e47d57003e6977ff9d73367d487