Introduce support for SoT in large repo
Summary:
## This stack
Implements the logic on `git pushrebase` to handle pushes when the source of truth is in the large repo (e.g. fbsource).

## This diff

Introduces the scaffolding (e.g. new modules, entrypoints) for the new flow (i.e. SoT in large repo) and the logic to determine when this flow should be used.

FYI, the actual logic of `get_large_repo_name` will be in an SCS command that I haven't yet implemented. But I didn't want to block all the work on that, so for now I set up a way to mock its result through environment variables.

This let me set up almost the entire logic and start writing some integration tests.

Reviewed By: mitrandir77

Differential Revision: D62742807

fbshipit-source-id: 42296cf8c528cc2f250ac3fbefd8124d9e78a306