Add integration test for git pushrebase with Mononoke Git and SCS
Summary:
## This stack
Implements the logic on `git pushrebase` to handle pushes when the source of truth is in the large repo (e.g. fbsource).

## This diff
Adds the first git-pushrebase integration test for when the repo's SoT is in the large repo.

I purposely extracted the entire set up into a bash helper function, because, unlike my previous integration tests which have been long, I plan on setting up multiple small integration tests for git-pushrebase.
This is because it's a little more complicated to write unit tests for git pushrebase, since it spawns many shell commands.

I can do that by adding an intermediate interface for `Command`  instances, which I can mock, but this will take some time and I'm not sure if it's worth it if I can just set up integration tests.

Reviewed By: mitrandir77

Differential Revision: D62574976

fbshipit-source-id: 8c8325ad861ca0e573c700d4e59fbd15428204e7