Fix: Add stable flag to hg debug network speed to control runtimes
Summary:
# Bug Description
Users have reported doctor reporting network speed timeouts when running doctor when their network is fine.

# Cause
Hg debugnetwork --speed has a feature where if the initial test succeeds under 200ms, it is run again at 25x the filesize. On a bursty network, this can cause the succeeding runs to take upwards of 15 seconds to complete.

# Fix
Add a stable flag to hg debugnetwork to not run the 25x size increase

Differential Revision: D62889246

fbshipit-source-id: 6ee7c2ef624e909c8e6c15b7b403a001d955904a