Rename and expose util functions that get sumodule deps
Summary:
# This stack
Setting up a robust integration test to test Mononoke Git server with xrepo infra and submodule expansion.
This will test switching the source of truth from Mononoke Git to large repo.

## This diff
Rename a couple of util functions and make one of them public./
Currently, only the one that gets for both source and target repo pair is public.

In the next diff, we'll use the one that gets for a single repo directly.

Reviewed By: mitrandir77

Differential Revision: D62293107

fbshipit-source-id: 0f8f99640811780937477d6463d5bd7079217682