Refactor: move the supported repos to the constructor
Summary:
## This stack

Convert one (non-Megarepo) SCS method to run via async_requests and the async worker.

## This diff

When constructing a queue client we already know what repos we want to handle; there is no point in passing the list every single time. See later in the stack for how this is useful.

Reviewed By: singhsrb

Differential Revision: D62870848

fbshipit-source-id: a3b857454d3d503da8fe08f54f4308a0030795cd