metalog: add slots to track extra git references
Summary:
See the previous commit for the problem. We need to update Git refs during
commit rewrite. To do that, we need to first track what Git refs are there.
While we can read directly from Git, it's cleaner to go through metalog so
the transaction, locking etc work the same as other cases.

Reviewed By: muirdm

Differential Revision: D62786535

fbshipit-source-id: 00ed935a600b7deada49c282686255e9664aedbb