test-dotgit-visibleheads: demonstrate stale "main" branch pins obsolete visible
Summary:
Took me some time to figure out this repro. We turn local Git branch "main" to
a visible head. A side effect of that is that rewriting the commit no longer
updates the Git ref and the Git ref will keep the commit alive. This diff adds
a test demostrating the issue.

Reviewed By: sggutier

Differential Revision: D62782294

fbshipit-source-id: a2fb5aa23e9cb3c4d548bd61c0c7ee9a6e7a0345