metalog: use system time for default commit time
Summary:
See the pevious commit for context. The metalog API's
`CommitOptions::default()` produces the "0" timestamp
on commit. Move the "now" logic from pymetalog to metalog so
`CommitOptions::default()` can default to use the current clock.

This provides more meaningful timestamps for debugging, and makes it hard to
produce conflicting hashes.

Reviewed By: muirdm

Differential Revision: D62779907

fbshipit-source-id: c901c6c7e4e721200e06dcf46f12ebbd8e51850d