Subscribe to current diff and fetch the diff's comments
Summary: use the new `VSCodeReposList.observeActiveRepos` to listen for all the mounted repos. Then, watch the head (`.`) commit as it changes, and use the repo's CodeReviewProvider to watch for diff summary fetches.  Use both of these event streams to trigger updates for what comments are active for this repo. In the following diffs, this will add gutter actions to show comments for the relevant files.

Reviewed By: quark-zju

Differential Revision: D62540422

fbshipit-source-id: bd099c999ccc9ac2bf494ecc468df34b444216b7