Add a JK so we can safely roll out the new params getter
Summary:
## This stack

Convert one (non-Megarepo) SCS method to run via async_requests and the async worker.

## This diff

The change in the previous diff is somewhat risky; protect it with a JK.

I tried writing a helper function or macro but it led me to generics hell. The copypasta is unfortunate but will be short-lived anyway.

Reviewed By: mitrandir77

Differential Revision: D63013826

fbshipit-source-id: 7ee0bb55b4592763b12effca5604507401a97445