Read the repo id for ACL checks from the params
Summary:
## This stack

Convert one (non-Megarepo) SCS method to run via async_requests and the async worker.

## This diff

The megarepo commands are performing ACL checks before scheduling the job, and then again on poll. That makes sense for defense in depth, though it might be overkill.

Currently that is implemented by round-tripping the original `target` args in the token: SCS copies the target into the token, and expects the client to send it back unchanged. That's not particularly secure, really, because there is no signature on the token, and no cross-check that the target matches the job.

Unfortunately it also gets in the way of making this code more flexible: the current code depends on the `target` pointing to a repo and probably also a bookmark. Making it generic enough to support all kinds of methods that take only either (or none) would be a lot of work and make the code needlessly complex.

Instead, we can just read back the original params corresponding for the request and check. That is arguably more secure, at the price of one more DB access. At the current qps (measured in dozens of requests per hour), that is hardly a performance issue. Should it become one, we can figure out caching.

Reviewed By: mitrandir77

Differential Revision: D62971471

fbshipit-source-id: 73a09ac29f15587526e47faf1b3ad9321163bf20