case_conflict_skeleton_manifest: add odd and even depth conflict rollup counts
Summary:
Adds two more rollup counts to `CaseConflictSkeletonManifest`: `odd_depth_conflicts` is the number of descendants that have more than one child and have odd depth (i.e. children or 3rd level descendants or ...). Similarily, `even_depth_conflits` is the number of descendants that have more than one child and have even depth (i.e. the manifest itself or 2nd level descendants or ...).

The purpose of these rollup counts is to be able to tell if a manifest has case conflicts without traversing it, because for the root `CaseConflictSkeletonManifest` the `odd_depth_conflicts` will indicate the number of case conflicts that the manifest has.

Reviewed By: singhsrb

Differential Revision: D63092870

fbshipit-source-id: a34040c2807842ab9809e17e2f8c7e4a73c428f5