Add dynamic counter for Overlay size per repo
Summary:
## Context:
Previously, we added counters for all the operations in Overlay. However, we don't have any counter to show the size of the files and trees in Overlay. Like what we have for `inode`, we can add new dynamic counters are report size of Overlay per mounted repo.

## This diff:
Add new `overlaySize_` that has `dirCount` and ++ and -- the value when a tree enter or remove from Overlay

Reviewed By: jdelliot

Differential Revision: D62781586

fbshipit-source-id: 6ca66bebac7b4d2035aa0c8b0fc6aeedc68a1174