Add new counter for fileCount to `overlaySize_`
Summary:
## Context:
Previously, we added counters for all the operations in Overlay. However, we don't have any counter to show the size of the files and trees in Overlay. Like what we have for `inode`, we can add new dynamic counters are report size of Overlay per mounted repo.

## This diff:
Add `fileCount` and ++ and -- the value when a file enter or remove from Overlay

Reviewed By: MichaelCuevas

Differential Revision: D62913560

fbshipit-source-id: 35b4b903c655e5910506f6e4a668b8c236373bd3