Add option to preferred code review for branching PRs
Summary:
Add a new option to "preferredSubmitCommand" to control the branching PR workflow as "sl push". This way, it's another alternative to `sl pr` and `sl ghstack`. This will also encompass pushing with creating PRs, which I consider the same sort of thing. We'll update the UI to say "push" instead of "submit" when using this mode.

I could also imagine a separate setting for "branch" vs "stacked", and then use the preferred submit command only if using "stacked". But since they're mutually exclusive, why not consolidate and use the existing setting?

Reviewed By: muirdm

Differential Revision: D61569914

fbshipit-source-id: 9b7e0d58eff87dfd4de372e0ac8cf74fd78b395a