Show "Push..." instead of submit if using branching PR flow
Summary:
Start using the `push`  option for the preferred submit command setting to control the commit info view "submit" button. Here we change it to say "push..." instead, to indicate that it will be doing a push instead of a "submit", and the "..." indicates that more interaction will be required.

In the following diffs we'll show the Push details modal when you click push. This is where you can choose where to push to (branch name, etc), and create a PR if desired.

In effect, the modal will return the resulting queueable commands to do the `sl push` and then (possibly) create the PR (probably with `gh pr create`).

Reviewed By: muirdm

Differential Revision: D61569913

fbshipit-source-id: 3bc5ad4dd8a113dbe59854d4a2ea4d27c2c0a956