Run sl push on confirm
Summary:
Add a PushOperation and hook it up to the Push button in the modal. For now, this does not include different destinations. It passes the remote branch name and the commit to push.

After this, we can examine the output to auto-open GitHub to create the PR, or use a `gh` command we queue up to make the PR without going to github.

Reviewed By: muirdm

Differential Revision: D61948401

fbshipit-source-id: 3d5f57058ba7ea120f5a308ed1f5eb6272fc5598