Convert remote bookmark into equivalent fetched branch name
Summary:
Now that sl in transparent git mode is able to see the git branches as remote bookmarks, we can show the diff badges for branching PRs.

However, this requires a translation step:
remote branches are like "origin/feature/myPr" but the "branchName" we get from github via graphql is just "feature/myPr".

We need to do that translation somewhere. I opt to have the code review provider do this. I think "origin/" will be the most common prefix, but I don't think it has to be "origin"...? We should probably look this up what the remote name is using `sl config`.

Assuming we still want to use the remote bookmark as the definining feature for the PR.

Reviewed By: quark-zju

Differential Revision: D62414975

fbshipit-source-id: 8079b6cbbf20f3f8f58ab7ef41b794f25efd8908