Index diff infos by branchName and show badge for matching bookmarks
Summary:
In order to show branching PRs, we need to know the appropriate branching PR for a given commit stack.

Our plan is to render branching PRs on the headRef (which is where the remote bookmark is).

To do this, we can construct a mapping of branch name to diffID. Then our normal diff info lookup works like normal.

This lets us render branching Prs on commits, by hijacking the commit's remote bookmarks and rendering each of them as a (<Bookmark>, <DiffBadge>) pair. This will be expanded in the future with more features of course. This just gets things started by showing PR badges by branch name.

note that public commits still need to show their remote bookmarks, but not as PRs. So they skip this code path and still render their remote bookmarks among the other bookmarks.

Reviewed By: quark-zju

Differential Revision: D61503759

fbshipit-source-id: f4d79a2072d849214cd81783272a1b46affc5857