Add field for branch name
Summary:
When you push, you need to pick a destination to push to. When making a PR, you'd push to a new remote branch that gets created in the process.

We need to ask what branch they'd like to use. To seed it, we can auto-gen a branch name from your stack. For now, this is just `feature/top-commit-title` with some sanitization. Eventually this should be configurable. What you type in here should also get validated for format.

I think it would be nice if the UI made it clear if you're pushing to an existing branch name or not. We should look up what you've typed in there and show "creating new branch named" or "updating existing branch named".

Reviewed By: muirdm

Differential Revision: D61947931

fbshipit-source-id: 4cef47fffa7229c04c5825bf6a157b7490bb46ce