Pass github provider to PR modal
Summary:
Validate the UICodeReviewProvider is a github provider, and pass it to the PR modal. This can be used to find the default push location and do other github-specific stuff.

I'm still consider how much of this PR modal flow should live inside github-specific code and how much we can make be reusable. I think most of this should not be in UICodeReviewProvider, so in the future something like GitLab support could be added in a very similar way.

Reviewed By: muirdm

Differential Revision: D61946174

fbshipit-source-id: 79ada54ed1e04b8dcc8a3f29ed68f7aeef0f952e