mergetool: add a new conflict marker format 'mergediff2'
Summary:
This conflict marker format is based on `'mergediff`, and it clearly shows the difference between the base content and the content from each side of the merge, which will be very useful to investigate if the merge algorithms works as expected.

I also often find this format helps to resolve conflicts. Since when I resolving conflicts, I often need to understand the changes of both sides. This is also similar to the `bk smerge -g` output format.

The new format is like below, the `<<<<<<< working copy` section is the diff between `source` and `base` commits, and the `>>>>>>> destination` section is the diff between `dest` and `base` commits:

```
  <<<<<<< working copy
  -3
  +3.5
   4
   5
  ======= base
   3
  -4
  -5
  +6
  +8
  >>>>>>> destination
```

Later, we can also minimize the diff output by moving the same changes of both sides out of the conflict area. Similar to `zdiff3` style from Git, or our default `minimized` format.

ignore-conflict-markers

Reviewed By: muirdm

Differential Revision: D62336747

fbshipit-source-id: d34f33f45c29cc2e3eebcaabc31bc591be4669e2