Stop enforcing write permissions on poll methods
Summary:
## Background

D28838994 originally added ACL checks to all Megarepo methods. `*_poll` methods don't receive all params, so instead the code depended on the `token` carrying enough information (the "target") for the ACL check.

That had cascading consequences that made the code complex when trying to accommodate new methods that may not have a target repo and/or bookmark. It also added no security, because the client could easily replace the target repo id in the token with one they controlled (the token includes no signature).

Moreover, there was never a requirement to enforce write permissions, given that `*_poll` methods don't write any new data nor does it disclose anything that can't be observed by accessing the repo. Concerns on the usefulness of these checks were raised immediately on D28838994 but until now they didn't hurt so they stayed in.

## What changed

D62971471 changed the code to stop assuming the token carried a token, and instead retrieving the params from the original request. Each `*_poll` method knows to expect the right params, so it is easier to accommodate different arguments and possibly different permissions. This changed passed integration tests and was merged (with a JK added in the next diff, D63013826).

In production this worked most of the time, but once in a while (e.g. I observed 2 instances in 2 hours) this code fails to find the token. This was the root cause of S451575.

The issue is that this code is racy by design. When the client calls the first method, data gets inserted into a DB and a token returned to the client. When the client then makes a call to the corresponding `_poll` method, the DB change may or may not have replicated. The core logic of `poll` simply retries if the entry is missing from the DB; but the initial check doesn't.

## This diff

This diff removes the ACL check and all supporting code, undoing most of D28838994.

Reviewed By: singhsrb

Differential Revision: D63310843

fbshipit-source-id: 1945a15a794b780466373cb130cdccf789748166