Add short-hand configuration option in Mononoke App for Git repos
Summary: There are a lot of `monad` commands (and other utilities) that we use just for Git repos. This requires us to specify the configuration path or tier for Git repos with `--config-tier gitimport_content`. This diff provides a short hand in the form of `--git` that serves the same purpose.

Reviewed By: YousefSalama

Differential Revision: D63324277

fbshipit-source-id: 0badc973a94755551d6201afa0bf0a0689467a3e