add more info to mount XCHECK
Summary:
# This diff

Adds some additional debugging info to an XCHECK that is consistently causing crashes (~100 per month).

# Context

This XCHECK ensures that a previous mount operation hasn't already added a given mount point to the mount map. If it has, then we crash Eden. The issue with this is that we don't have any information about why this check is failing sometimes. One running theory is that the lifetime of the path piece could be wrong and therefore we're attempting to insert empty strings into the mount map.

To test this theory, this diff adds some additional info about the duplicate path to the mount map.

Reviewed By: genevievehelsel

Differential Revision: D63293234

fbshipit-source-id: cbad75a1ac863dd6588eaef2f2999c08ea1970d0