backingstore: try to reduce window for lost cache writes
Summary:
When we hot-reload the backingstore, we may lose cache writes since we don't perform any locking when swapping the stores. For example:

1. Flush old store
2. Make new store
3. Swap in new store as active store

If we make new cache writes between 1. and 3. to the old store, they might be lost after swapping the new store. Attempt to reduce the window by flushing the old stores again after 3.

Reviewed By: quark-zju

Differential Revision: D63353607

fbshipit-source-id: b21d116d8e9812f338d946352a7f713fd444f4a9