serialize sync target config and put it in the aosp megarepo working copy
Summary:
more context for why in this doc: https://fburl.com/gdoc/una2ler8

Let's serialize the current config and put it into the repo. To keep the config file up-to-date we need to do this in two scenarios:
 * when adding new target.
 * when changing config version for a target

Note that we don't need to do it during sync_changeset or remerge_source calls as those don't change the version we're on.

Reviewed By: cfsmp3

Differential Revision: D63316789

fbshipit-source-id: b48c8b03296739dbd18626641a020409ae5f9d7d