mononoke_api: make ChangesetPathContext independent of SkeletonManifest
Summary: `ChangesetPathContext` stores a `SkeletonManifest` entry but only cares about whether the entry exists and whether it's a tree or a leaf. This diff makes it store an `Option<Entry<(), ()>>` instead of `Option<Entry<SkeletonManifestId, ()>>` to make it independent from `SkeletonManifestV1`. This will make it easier to migrate to `SkeletonManifestV2`

Reviewed By: lmvasquezg

Differential Revision: D63318353

fbshipit-source-id: 4b5cc7c54ff3cd17851f049202e145bf74188ec0