Stop assuming all methods have a bookmark parameter
Summary:
## This stack

Convert one (non-Megarepo) SCS method to run via async_requests and the async worker.

## This diff

Existing methods will still be able to refer to their target bookmark via the params. The new methods we are adding won't refer to any bookmark at all.

We were not using the bookmark anywhere, except for the CLI.

This diff can only land after changing the DB to make the column nullable. We can drop it later on.

Reviewed By: gustavoavena

Differential Revision: D63081986

fbshipit-source-id: a27353ed0cc5964a10dee39310a0fdf2c410c927