Show known diagnostics before commit/amend
Summary:
Before committing/amending, we can warn you about known issues in the files you're touching.

This is "best effort", since vscode only knows these diagnostics for files that are open and have had the LSP run correctly.

This diff opts to ask vscode for diagnostics for files that are in your list of uncommitted changes that are selected. There may be files with errors caused by your change that are not included in this set.

It's also possible to still cause issues in commits that bypasses this check, for example by splitting or using edit stack to make commits. The check is not run on "submit" either, just on amend/commit (though we could add the check during submit for files in the current commit). But submit may be run on commits that are not `.`.

Regardless, we intend for CI to be the more authoritative check for errors. This is meant to be a first line of defense, to quickly let you know about errors when possible.

Some other details:
- As mentioned, this only asks for diagnostics for selected files to commit/amend
- We only block if there are `error` level diagnostics
- We show all the found diagnostics for the files, even non-error ones
- TODO: this should run on the quick commit/amend buttons, not just the commit info view ones
- TODO: you should be able to click diagnostics to open them
- TODO: we should have a config for this, so that we can add a "don't show again" checkbox at the bottom of this confirmation

Reviewed By: quark-zju

Differential Revision: D62788210

fbshipit-source-id: 2ef8b9012e415eaee5697e366a974acad55d2596