Only open diagnostic confirmation for submit actions
Summary:
Pivot slightly from previous approach. Instead of showing the diagnostics confirmation anytime you commit/amend, we'll show it (in certain cases) when you _submit_.

The most common way to submit is through "Commit & Submit" and "Amend & Submit", which means it will still show up in most cases. It will no longer appear when doing a "quick commit".

This will reduce cases of ignoring errors when doing a quick commit, then having to ignore them AGAIN to submit. Granted, that's still possible if you resubmit the commit again later with the same errors...

Unfortunately, we can't use the diagnostics confirmation when submitting commits you're not checked out on. We can't even confirm about errors when submitting the head commit if you have uncommitted changes to the same file. Maybe we can do something about that in the future.

Reviewed By: muirdm

Differential Revision: D63423865

fbshipit-source-id: 815d3a39224dd222e86a5c92c29f0512eab1d2f0