Add checkbox to hide non-blocking diagnostics
Summary: We get all diagnostics from vscode, but only 'error' severity issues really matter. Let's hide the other ones by default, but give you a checkbox to show them all (in case you get value from the linter suggestions). We use persistant storage for this so your choice is remembered (but no need for a slot in the settings dropdown for this)

Reviewed By: quark-zju

Differential Revision: D62960882

fbshipit-source-id: 6543ae2535f0cb824f324e1be2b28f84ffb61851