Add support to fetch vscode diagnostics
Summary:
Add types for vscode diagnostics to be collected on the server and sent to the webview. These can be used in a pre-commit or pre-amend confirmation modal.

Currently, this only sends requested files. We could ask for all diagnostics and filter by repo—in case it's useful to get information about files you're not committing but had open and got diagnostics in....?

Reviewed By: quark-zju

Differential Revision: D62788209

fbshipit-source-id: a7ea301179f9c0b4a276975cf4172be363e13f4e