Add tracking for diagnostics
Summary:
Add track calls to the diagnostics confirmation

We want to be able to answer questions like:
- Which languages typically have errors that would be shown by the confirmation?
- Who saw the confirmation?
- Who could have seen the confirmation, but was excluded by the experiment?
- Who could have seen the confirmation, but did not have any applicable errors?
- How many errors do people typically have?
- Do people typically have non-error warnings?

We accomplish this with a track call before the confirm is shown, which happens regardless of if you would get the confirmation. This includes the sources of errors and count. We have a second track call that logs if you were shown the modal, and says how you responded: continue (ignore) or cancel. I think these two calls should let us answer all the above questions.

Reviewed By: muirdm

Differential Revision: D63421610

fbshipit-source-id: 7fb4d3b25cb3b1c6b843921ea0b0e06da05c3a71