Expose promise based feature flag API
Summary:
Previously, feature flags were implemented with `lazyAtom`. This made it hard to get at the underlying atom to cache the promise so you could await a fetch without going through the UI.

It's arguably simpler if we don't use lazyAtom at all and just have the atom be an `Atom<Promise<boolean>>`, and then use `loadable` to get at it synchronously.

In my testing, this doesn't make it overfetch the GKs.

Reviewed By: muirdm

Differential Revision: D63416529

fbshipit-source-id: 54e7cac21de2b33b20a7e5b6932b8c6208353751