Add a new type to report errors in async methods
Summary:
## This stack

Convert one (non-Megarepo) SCS method to run via async_requests and the async worker.

## This diff

Megarepo-related methods use a very specific Thrift union to report errors, which is unfortunate to use in generic code. Renaming it will need cross-teeam coordination, it can wait.

Instead I'm adding a new generic type that we can use for new methods. When that has stabilized we can figure out the migration for old methods (or not).

Reviewed By: RajivTS

Differential Revision: D63251903

fbshipit-source-id: 3979da03c3969163e27b179fffd598d2f5f79e7d