Simplify the target method to be display only
Summary:
## This stack

Convert one (non-Megarepo) SCS method to run via async_requests and the async worker.

## This diff

The `target()` method is now unused, except for two kinds of human consumption purposes:

* in the CLI (i.e. `monad async-requests list`);
* in the worker logs

Both of these are useful enough to preserve.

Ideally we'd implement `Display`, but doing that for Thrift-generated classes is a bit more involved than I have time for right now. I'll do it later, but in the meantime this unblocks us to add new methods.

Reviewed By: RajivTS

Differential Revision: D63136539

fbshipit-source-id: 2ca58516e219d4b1cba9cf913f8e60ec12fc1414